% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saalfeld_reg.R
\name{add_saalfeldlab_reglist}
\alias{add_saalfeldlab_reglist}
\title{Register single ANTs registration folder}
\usage{
add_saalfeldlab_reglist(x, ...)
}
\arguments{
\item{x}{Path to a single ANTs registration folder named according to
Saalfeld conventions}

\item{...}{Additional arguments passed to \code{\link{add_reglist}}}
}
\description{
Register single ANTs registration folder
}
