% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saalfeld_reg.R
\name{download_saalfeldlab_registrations}
\alias{download_saalfeldlab_registrations}
\title{Download Saalfeld Lab registrations}
\usage{
download_saalfeldlab_registrations(fileformat = c(".h5", ".nii"))
}
\arguments{
\item{fileformat}{whether to download h5 (Saalfeld) or nii (ANTs) format
registrations (defaults to h5).}
}
\description{
Download Saalfeld Lab registrations
}
\details{
Registrations will be downloaded from
  \url{https://www.janelia.org/open-science/jrc-2018-brain-templates}. They
  will be downloaded to a folder defined by the package option
  \code{nat.templatebrains.regdirs} (see examples).

  The h5 registration format typically provides a multi-resolution forward
  and inverse transformations in an efficiently compressed format. See
  \url{https://github.com/saalfeldlab/template-building/wiki/Hdf5-Deformation-fields}
   for further details.

  The nii (NIfTI) files are the raw output format of the ANTs registration
  package. Their disadvantages are that they are 1) large files and the whole
  >500 Mb registration file must be loaded completely into memory before any
  transformations are possible, 2) that they depend on a more complex set of
  packages that can be hard to install, 3) is not simple to downsample to
  make smaller files. They do have the advantage of support for transforming
  image data as well as 3D points.
}
\examples{
regroot=getOption('nat.templatebrains.regdirs')
dir(regroot)
}
