% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nat.jrcbrains-package.R
\docType{package}
\name{nat.jrcbrains-package}
\alias{nat.jrcbrains}
\alias{nat.jrcbrains-package}
\title{nat.jrcbrains: Janelia Research Campus Bridging Registrations for Drosophila Brains}
\description{
The main purpose of this package is to distribute bridging
  registrations mapping between template brains constructed at Janelia
  Research Campus by Stephan Saalfeld's group. These registrations are not
  directly distributed with the package but must be downloaded. See
  \code{\link{download_saalfeldlab_registrations}} for details.

  Once the appropriate registrations have been downloaded, loading the
  package is sufficient to make them available to end users / other packages.
  See \code{\link{register_saalfeldlab_registrations}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/natverse/nat.jrcbrains}
  \item Report bugs at \url{https://github.com/natverse/nat.jrcbrains/issues}
}

}
\author{
\strong{Maintainer}: Gregory Jefferis \email{jefferis@gmail.com} (\href{https://orcid.org/0000-0002-0587-9355}{ORCID})

}
\keyword{internal}
