% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saalfeld_reg.R
\name{register_saalfeldlab_registrations}
\alias{register_saalfeldlab_registrations}
\title{Register Saalfeld Lab registrations with nat.templatebrains}
\usage{

  register_saalfeldlab_registrations(x = getOption("nat.jrcbrains.regfolder"),
  ...)
}
\arguments{
\item{x}{A folder containing downloaded Saalfeld lab registrations}

\item{...}{Additional arguments passed to
\code{\link{add_saalfeldlab_reglist}} and eventually to
\code{\link{add_reglist}}}
}
\value{
A named logical vector indicating which of the subfolders was
  successfully registered with \code{\link{nat.templatebrains}}
}
\description{
Register Saalfeld Lab registrations with nat.templatebrains
}
\details{
You must download the registrations from
  \url{https://www.janelia.org/open-science/jrc-2018-brain-templates} using
  \code{\link{download_saalfeldlab_registrations}}. If you prefer you may
  also download them manually to to a single folder, setting
  \code{options(nat.jrcbrains.regfolder)} to point to this folder in your
  \code{\link{Rprofile}}.
}
\examples{
\dontrun{
options(nat.jrcbrains.regfolder='/GD/projects/JFRC/JohnBogovic/jrc-2018-brain-templates')
register_saalfeldlab_registrations()
}
}
\references{
See Bogovic et al. (2018) \url{https://doi.org/10.1101/376384}
}
\seealso{
\code{\link{download_saalfeldlab_registrations}}
}
